move jwk interface to webcontainer proj
