+++ dev/com.ibm.ws.security.jwt/src/com/ibm/ws/security/jwt/utils/JwtData.java
@@ -20,10 +20,10 @@ import com.ibm.websphere.ras.TraceComponent;
 import com.ibm.websphere.ras.annotation.Sensitive;
 import com.ibm.ws.ffdc.annotation.FFDCIgnore;
 import com.ibm.ws.security.common.jwk.impl.JWKProvider;
-import com.ibm.ws.security.common.jwk.interfaces.JSONWebKey;
 import com.ibm.ws.security.jwt.config.JwtConfig;
 import com.ibm.ws.security.jwt.internal.BuilderImpl;
 import com.ibm.ws.security.jwt.internal.JwtTokenException;
+import com.ibm.ws.webcontainer.security.jwk.JSONWebKey;
 
 /**
  *
