+++ dev/com.ibm.ws.webcontainer.security/src/com/ibm/ws/webcontainer/security/jwk/package-info.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2016 IBM Corporation and others.
+ * Copyright (c) 2018 IBM Corporation and others.
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
  * which accompanies this distribution, and is available at
@@ -8,20 +8,9 @@
  * Contributors:
  *     IBM Corporation - initial API and implementation
  *******************************************************************************/
-package com.ibm.ws.security.common.jwk.interfaces;
-
-import com.ibm.json.java.JSONObject;
-import com.ibm.ws.webcontainer.security.jwk.JSONWebKey;
 
 /**
- * TODO: This should replace com.ibm.ws.security.openidconnect.jwk.JWK
+ * @version 1.0
  */
-public interface JWK extends JSONWebKey {
-
-    public void parse();
-
-    public void generateKey();
-
-    public JSONObject getJsonObject();
-
-}
+@org.osgi.annotation.versioning.Version("1.0")
+package com.ibm.ws.webcontainer.security.jwk;
